s/sparkle/brain/g
